+++ src/test/scala/priority_queue/MaxPriorityQueueTest.scala
@@ -72,7 +72,7 @@ class MaxPriorityQueueTest extends FunSuite with BeforeAndAfter {
   }
 
   test("addition of a lot of elements causes extension of a heap") {
-    1 to 20 foreach (x => pq.insert(x.toDouble))
+    1 to 20 foreach (x => pq.insert(x.toInt))
 
     assertResult(20) {
       pq.deleteMax()
